fix: 修复 QueryAwareAdapter 的 dtype 不匹配问题
mask.float() 会转换为 float32，但模型权重是 bfloat16。
改为 mask.to(dtype=dtype) 以匹配模型的 dtype。

+++ method/models/adapter.py
@@ -153,7 +153,8 @@ class QueryAwareAdapter(nn.Module):
 
         # 1. Encode mask pattern
         # mask: (batch, n_vision) -> (batch, hidden)
-        mask_float = mask.float()
+        # 确保 mask 的 dtype 与模型权重一致
+        mask_float = mask.to(dtype=dtype)
         mask_emb = self.mask_encoder(mask_float)
 
         # 2. Pool question into single vector
